FIX BUG: jsdi_jni.cpp -- functions templated on InvokeFunc not always calling invokeFunc Rationalize exceptions framework: there should now be JNI exception checks at every location which could trigger a JNI exception
NOTE: If you run the JVM with -Xcheck:jni, it still reports the following warning: 'WARNING in native method: JNI call made with exception pending'. This seems to be because some of the jni_util auto-objects call JNI functions in their destructors. I don't *think* the specific instances called in the destructors should cause a problem, as long we stop executing non-stack-unwind code at the instant an exception is flagged to JNI.