FIX BUG: Remove trailing newline from PrintStdout().
My previous change added a call to System.out.println() in order to add a trailing newline to anything printed with PrintStdout(). After discussing with APM, this extra newline should be removed because PrintStdout is intended to be a Suneido.Print function used by stdlib Print() and stdlib Print() already adds a trailing newline.